% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refClass_summaryData.R
\docType{class}
\name{SummaryData-class}
\alias{SummaryData-class}
\alias{show,SummaryData-method}
\alias{SummaryData}
\title{Class SummaryData}
\usage{
\S4method{initialize}{SummaryData}(.Object, data = "data.frame",
  data_name = "character", summary_columns = "character",
  info = "character", ...)

\S4method{show}{SummaryData}(object)

SummaryData(...)
}
\value{
A SummaryData Class
}
\description{
Odd class facilitating creation of data frame with
  summary information attached.

Show method
}
\section{Slots}{

\describe{
\item{\code{data}}{data.frame.}

\item{\code{summary_columns}}{character.}

\item{\code{info}}{character.}
}}

