% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refClass_summaryData.R
\docType{class}
\name{SummaryData-class}
\alias{SummaryData-class}
\title{Class SummaryData}
\value{
A SummaryData Class
}
\description{
Odd class facilitating creation of data frame with
  summary information.
}
\section{Slots}{

\describe{
\item{\code{data}}{data.frame.}

\item{\code{summary_columns}}{character.}

\item{\code{comment}}{character.}
}}

